+++ src/Frontend/Dfe.Complete/Extensions/ClaimsPrincipalExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System.Security.Claims;
+using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Application.Projects.Queries.GetUser;
 using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Utils;
@@ -25,6 +26,29 @@ namespace Dfe.Complete.Extensions
         {
             return user?.Claims.Any(c =>
                 c.Type == ClaimTypes.Role && string.Equals(c.Value, role, StringComparison.OrdinalIgnoreCase)) ?? false;
+        }
+        
+        public static async Task<UserDto> GetUser(this ClaimsPrincipal value, ISender sender)
+        {
+            try
+            {
+                var userAdId = value.GetUserAdId();
+
+                var request = new GetUserByAdIdQuery(userAdId);
+                var userResult = await sender.Send(request);
+
+                if (!userResult.IsSuccess || userResult.Value == null)
+                {
+                    throw new NotFoundException(userResult.Error ?? "User not found.");
+                }
+
+                return userResult.Value;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
+            }
         }
     }
 }